# openapi.yaml
openapi: 3.1.0
info:
  title: LearnFast Core Engine
  description: Hybrid Graph-RAG Learning Platform API
  version: 0.1.0
paths:
  /api/documents/upload:
    post:
      tags:
      - documents
      summary: Upload Document
      description: "Uploads a new study document and extracts its content.\n\nProcesses\
        \ the file (PDF or Image), saves it to disk (via DocumentStore), \nand uses\
        \ the DocumentProcessor to extract text."
      operationId: upload_document_api_documents_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_documents_upload_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/youtube:
    post:
      tags:
      - documents
      summary: Ingest transcripts from a YouTube video
      description: 'Ingest transcripts from a YouTube URL.

        Fetches transcript, saves as a virtual document, and processes it.'
      operationId: ingest_youtube_api_documents_youtube_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_ingest_youtube_api_documents_youtube_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/:
    get:
      tags:
      - documents
      summary: Get Documents
      description: Retrieves a list of documents with optional filtering.
      operationId: get_documents_api_documents__get
      parameters:
      - name: folder_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
                title: Response Get Documents Api Documents  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/{document_id}:
    get:
      tags:
      - documents
      summary: Get Document
      description: Retrieves a specific document's metadata.
      operationId: get_document_api_documents__document_id__get
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - documents
      summary: Update Document
      description: Updates document metadata (title, tags, category, folder).
      operationId: update_document_api_documents__document_id__put
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          title: Document Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - documents
      summary: Delete Document
      description: Deletes a document and its file.
      operationId: delete_document_api_documents__document_id__delete
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/{document_id}/move:
    put:
      tags:
      - documents
      summary: Move Document
      description: Moves a document to a different folder (or unfiles it).
      operationId: move_document_api_documents__document_id__move_put
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          title: Document Id
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_move_document_api_documents__document_id__move_put'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/{document_id}/session/start:
    post:
      tags:
      - documents
      summary: Start Reading Session
      description: Starts a time tracking session for a document.
      operationId: start_reading_session_api_documents__document_id__session_start_post
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/{document_id}/session/end:
    post:
      tags:
      - documents
      summary: End Reading Session
      description: Ends a time tracking session and updates progress.
      operationId: end_reading_session_api_documents__document_id__session_end_post
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          title: Document Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeTrackingRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_ingest_youtube_api_documents_youtube_post:
      properties:
        url:
          type: string
          title: Url
      type: object
      required:
      - url
      title: Body_ingest_youtube_api_documents_youtube_post
    Body_move_document_api_documents__document_id__move_put:
      properties:
        folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
      type: object
      title: Body_move_document_api_documents__document_id__move_put
    Body_upload_document_api_documents_upload_post:
      properties:
        file:
          type: string
          format: binary
          title: File
        title:
          type: string
          title: Title
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
          default: ''
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
      type: object
      required:
      - file
      - title
      title: Body_upload_document_api_documents_upload_post
...
    DocumentCreate:
      properties:
        title:
          type: string
          title: Title
        tags:
          items:
            type: string
          type: array
          title: Tags
          default: []
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
      type: object
      required:
      - title
      title: DocumentCreate
      description: Schema for creating a new document.
    DocumentResponse:
      properties:
        title:
          type: string
          title: Title
        tags:
          items:
            type: string
          type: array
          title: Tags
          default: []
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        id:
          type: integer
          title: Id
        file_type:
          $ref: '#/components/schemas/FileType'
          default: other
        file_path:
          type: string
          title: File Path
        upload_date:
          type: string
          format: date-time
          title: Upload Date
        extracted_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Extracted Text
        reading_progress:
          type: number
          title: Reading Progress
          default: 0.0
        folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
        time_spent_reading:
          type: integer
          title: Time Spent Reading
          default: 0
        last_opened:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Opened
        first_opened:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: First Opened
        completion_estimate:
          anyOf:
          - type: integer
          - type: 'null'
          title: Completion Estimate
        page_count:
          type: integer
          title: Page Count
          default: 0
      type: object
      required:
      - title
      - id
      - file_path
      - upload_date
      title: DocumentResponse
      description: Schema for document API responses.
    FileType:
      type: string
      enum:
      - pdf
      - image
      - other
      title: FileType
      description: Enum for supported document file types.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TimeTrackingRequest:
      properties:
        seconds_spent:
          type: integer
          title: Seconds Spent
        reading_progress:
          anyOf:
          - type: number
          - type: 'null'
          title: Reading Progress
      type: object
      required:
      - seconds_spent
      title: TimeTrackingRequest
      description: Request schema for updating document reading time.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError